添加基础功能

+++ lib/src/main/java/com/github/lib/bean/VideoClassify.java
@@ -15,8 +15,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class VideoClassify {
-    private Long page;
-    private Long totalPage;
+    private Integer page;
+    private Integer totalPage;
     private List<Video> videos;
     private List<Classify> classifies;
 }
